relocated the frame and canvas: now MainFrame.java
